[Feat] 둘다 채팅방에서 나간 경우 Storage에서 채팅 이미지 삭제

+++ BookBridge/BookBridge/ViewModel/Chat/ChatMessageViewModel.swift
@@ -637,21 +637,64 @@ extension ChatMessageViewModel {
 
 //MARK: 채팅방 나가기
 extension ChatMessageViewModel {
-    func deleteChatRoom(uid: String, completion: @escaping() -> ()) {
+    func deleteChatRoom(uid: String, partnerId: String, completion: @escaping() -> ()) {
         firestoreListener?.remove()
         
-        let messageRemoveQuery = FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(saveChatRoomId).collection("messages")
-        
-        messageRemoveQuery.getDocuments { querySnapshot, error in
+        FirebaseManager.shared.firestore.collection("User").document(partnerId).collection("chatRoomList").whereField("id", isEqualTo: self.saveChatRoomId).getDocuments { querySnapshot, error in
             guard error == nil else { return }
             guard let documents = querySnapshot?.documents else { return }
             
-            for document in documents {
-                messageRemoveQuery.document(document.documentID).delete()
-            }
+            let messageRemoveQuery = FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).collection("messages")
             
-            FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
-                completion()
+            //상대방도 채팅방이 없을 경우만 이미지 삭제
+            if documents.isEmpty {
+                messageRemoveQuery.getDocuments { querySnapshot, error in
+                    guard error == nil else { return }
+                    guard let documents = querySnapshot?.documents else { return }
+                    
+                    for document in documents {
+                        guard let imageURL = document.data()["imageURL"] as? String else { return }
+                        
+                        if imageURL != "" {
+                            self.deleteChatImages(imageURL: imageURL)
+                        }
+                        
+                        messageRemoveQuery.document(document.documentID).delete()
+                    }
+                    
+                    FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
+                        
+                        completion()
+                    }
+                }
+            } else {
+                messageRemoveQuery.getDocuments { querySnapshot, error in
+                    guard error == nil else { return }
+                    guard let documents = querySnapshot?.documents else { return }
+                    
+                    for document in documents {
+                        messageRemoveQuery.document(document.documentID).delete()
+                    }
+                    
+                    FirebaseManager.shared.firestore.collection("User").document(uid).collection("chatRoomList").document(self.saveChatRoomId).delete { _ in
+                        
+                        completion()
+                    }
+                }
+            }
+        }
+    }
+    
+    //채팅 이미지 삭제
+    func deleteChatImages(imageURL: String) {
+        //상대방도 채팅방이 없을 경우만 삭제
+        let ref = FirebaseManager.shared.storage.reference(forURL: imageURL)
+        
+        ref.delete { err in
+            if err == nil {
+                print("사진 삭제 성공")
+            } else {
+                print("사진 삭제 오류")
             }
         }
     }
